Separate states into files
