+++ src/Race/InitializedState.ts
@@ -1,0 +1,3 @@
+import { RaceState } from "./RaceState";
+
+export class InitializedState extends RaceState {}
